Fix loops
There were a problem with loops that I was keeping loop stack in the
main tree. This was problematic because code was executing outside of
loop and it created bugs in executing. Now I keep the instructions
inside the stack and it is pushed into a loop when needed. The match is
a bit yiky. Maybe will consider something better later